code/requirements.txt
Django==3
django-environ
django-extensions
django-cors-headers
psycopg2==2.8.6
djangorestframework==3.12.4
django-filter==2.4.0
